requirements.txt
astroid==2.4.2
certifi==2020.6.20
click==7.1.2
cmake==3.18.2.post1
cycler==0.10.0
decorator==4.4.2
distro==1.5.0
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
imageio==2.9.0
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
olefile==0.46
opencv-python==4.2.0.34
packaging==20.4
Pillow==7.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-build==0.11.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
